Added mouse event handlers
There are several raw and "generated" mouse event handlers that can be
hooked now. The "raw" ones are called on mouse input events and are
mouse_move, mouse_up, and mouse_down (up/down = pressed/unpressed). The
generated ones are mouse_click, mouse_enter, and mouse_exit (generated
from the other events as a convenience).

All of these can be overriden, but if so the super method should be
called to propogate the event and set proper variables (which the
generated events rely on).