Added radar and framework for feelers
The radar works, it draws an alpha-blended circle to show the radar
range, and turns detected enemies green.

Feelers do not work, but the framework for them is there.

To do radar, I made a Vector class. It's not complete, it only has
what's necessary, but it can be added to as need be (pun intended =).

I added some notes and a todo for Matt/Diego