Added logic for Feelers
Added a Line class to embody the logic for creating lines and finding
the intersection of lines.

Filled in the logic for the Feeler class. To test a feeler's
intersection with a wall, I find the intersection of the feeler's line
with the lines created by the points of a wall's rectangle. If the
intersection is within the rectangle (between the two points defining
the intersection line), then it's considered an intersection with the
wall (this is not yet done).

It finds the minimum distance to any of the valid intersections and this
is the first point that the feeler intersects a wall.

The difficulty is in line intersections because lines can be the same,
leading to infinite intersections, parallel, but not the same, leading
to zero intersections, or non-parallel and thus intersecting. In
addition, half the lines for the walls are vertical, meaning
intersections with vertical lines need to be considered and dealt with.