Finished feeler logic, cost function pseudo code
The feeler's logic should be finished, just have to draw them and add
some code to "remember" the intersections for the get_info function.
And, of course, they won't work the first time...

I also added pseudo code for the cost function to the Player class. It
doesn't get called and doesn't work, but the logic is correct, just have
to replace it with code that actually works, but does the same thing.